[Chore] #39 - 타입 변경으로 인한 관련 코드 수정

+++ src/main/java/org/moonshot/server/domain/keyresult/dto/request/KeyResultCreateRequestInfoDto.java
@@ -24,7 +24,7 @@ public record KeyResultCreateRequestInfoDto(
         Short idx,
         @NotNull(message = "KR 목표 수치를 입력해주세요.")
         @ValidTargetNumber
-        Integer target,
+        Long target,
         @NotNull(message = "KR 목표 수치의 단위를 입력해주세요.")
         String metric,
         @NotNull(message = "KR 목표의 이전 수식을 입력해주세요.")
